+++ example/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 from django_boost.models.mixins import JsonMixin
-from django_boost.models.fields import ColorCodeFiled
+from django_boost.models.fields import ColorCodeField
 
 from django_boost.models.mixins import (UUIDModelMixin, TimeStampModelMixin,
                                         LogicalDeletionMixin)
@@ -17,7 +17,7 @@ class Article(UUIDModelMixin, TimeStampModelMixin, LogicalDeletionMixin):
 class Tag(models.Model):
     name = models.CharField(max_length=64)
     category = models.ForeignKey(to="Category", on_delete=models.PROTECT)
-    color = ColorCodeFiled(upper=True)
+    color = ColorCodeField(upper=True)
 
 
 class Category(models.Model):
@@ -27,4 +27,4 @@ class Category(models.Model):
 class Customer(JsonMixin, models.Model):
     name = models.CharField(max_length=64)
     registered_at = models.DateField(auto_now_add=True)
-    color = ColorCodeFiled(upper=True)
+    color = ColorCodeField(upper=True)